chore: remove legacy deprecated dependencies and simplify build process (#9)
* circleci

* cleanup and rename package

* Update config.yml

* rename in circleci config

* version bump

* remove prod/dev differentiation in build process

* cleanup dependencies

* Update package.json

* fix yarn lock

* Update hardhat.config.ts


* Update hardhat.config.ts


* remove some lines

* remove lines in interfaceCompativilityTest.sol
