+++ contracts/tokenbridge/libraries/ProxyUtil.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /*
- * Copyright 2022, Offchain Labs, Inc.
+ * Copyright 2021, Offchain Labs, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,14 +16,16 @@
  * limitations under the License.
  */
 
-// solhint-disable-next-line compiler-version
-pragma solidity >=0.6.9 <0.9.0;
+pragma solidity ^0.6.11;
 
-import "../arbitrum/IArbToken.sol";
-import "../ethereum/ICustomToken.sol";
-import "../ethereum/gateway/IL1ArbitrumGateway.sol";
-import "../ethereum/gateway/IL1GatewayRouter.sol";
-import "../libraries/IWETH9.sol";
-import "../libraries/IERC165.sol";
-import "../libraries/gateway/ICustomGateway.sol";
-import "../libraries/gateway/ITokenGateway.sol";
+library ProxyUtil {
+    function getProxyAdmin() internal view returns (address admin) {
+        // https://github.com/OpenZeppelin/openzeppelin-contracts/blob/v3.4.0/contracts/proxy/TransparentUpgradeableProxy.sol#L48
+        // Storage slot with the admin of the proxy contract.
+        // This is the keccak-256 hash of "eip1967.proxy.admin" subtracted by 1, and is
+        bytes32 slot = 0xb53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d6103;
+        assembly {
+            admin := sload(slot)
+        }
+    }
+}
