+++ contracts/tokenbridge/libraries/Whitelist.sol
@@ -1,0 +1,76 @@
+// SPDX-License-Identifier: Apache-2.0
+
+/*
+ * Copyright 2021, Offchain Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+pragma solidity ^0.6.11;
+
+abstract contract WhitelistConsumer {
+    address public whitelist;
+
+    event WhitelistSourceUpdated(address newSource);
+
+    modifier onlyWhitelisted() {
+        if (whitelist != address(0)) {
+            require(Whitelist(whitelist).isAllowed(msg.sender), "NOT_WHITELISTED");
+        }
+        _;
+    }
+
+    function updateWhitelistSource(address newSource) external {
+        require(msg.sender == whitelist, "NOT_FROM_LIST");
+        whitelist = newSource;
+        emit WhitelistSourceUpdated(newSource);
+    }
+}
+
+contract Whitelist {
+    address public owner;
+    mapping(address => bool) public isAllowed;
+
+    event OwnerUpdated(address newOwner);
+    event WhitelistUpgraded(address newWhitelist, address[] targets);
+
+    constructor() public {
+        owner = msg.sender;
+    }
+
+    modifier onlyOwner() {
+        require(msg.sender == owner, "ONLY_OWNER");
+        _;
+    }
+
+    function setOwner(address newOwner) external onlyOwner {
+        owner = newOwner;
+        emit OwnerUpdated(newOwner);
+    }
+
+    function setWhitelist(address[] memory user, bool[] memory val) external onlyOwner {
+        require(user.length == val.length, "INVALID_INPUT");
+
+        for (uint256 i = 0; i < user.length; i++) {
+            isAllowed[user[i]] = val[i];
+        }
+    }
+
+    // set new whitelist to address(0) to disable whitelist
+    function triggerConsumers(address newWhitelist, address[] memory targets) external onlyOwner {
+        for (uint256 i = 0; i < targets.length; i++) {
+            WhitelistConsumer(targets[i]).updateWhitelistSource(newWhitelist);
+        }
+        emit WhitelistUpgraded(newWhitelist, targets);
+    }
+}
