--- hardhat.base-config.json
@@ -1,29 +1,0 @@
-{
-  "defaultNetwork": "hardhat",
-  "paths": {
-    "artifacts": "build/contracts"
-  },
-  "solidity": {
-    "compilers": [
-      {
-        "version": "0.6.11",
-        "settings": {
-          "optimizer": {
-            "enabled": true,
-            "runs": 100
-          }
-        }
-      },
-      {
-        "version": "0.8.7",
-        "settings": {
-          "optimizer": {
-            "enabled": true,
-            "runs": 100
-          }
-        }
-      }
-    ],
-    "overrides": {}
-  }
-}
