--- hardhat.dev-config.ts
@@ -1,201 +1,0 @@
-import '@typechain/hardhat'
-import '@nomiclabs/hardhat-waffle'
-import 'dotenv/config'
-import 'solidity-coverage'
-import 'hardhat-gas-reporter'
-import '@nomiclabs/hardhat-etherscan'
-import 'hardhat-deploy'
-
-import baseConfig from './hardhat.base-config.json'
-import { task } from 'hardhat/config'
-import '@nomiclabs/hardhat-ethers'
-
-task('accounts', 'Prints the list of accounts', async (taskArgs, bre) => {
-  const accounts = await bre.ethers.getSigners()
-
-  for (const account of accounts) {
-    console.log(await account.getAddress())
-  }
-})
-
-const config = {
-  ...baseConfig,
-  networks: {
-    hardhat: {
-      chainId: 1337,
-      throwOnTransactionFailures: true,
-      allowUnlimitedContractSize: true,
-      accounts: {
-        accountsBalance: '1000000000000000000000000000',
-      },
-      blockGasLimit: 200000000,
-      // mining: {
-      //   auto: false,
-      //   interval: 1000,
-      // },
-      forking: {
-        url: 'https://mainnet.infura.io/v3/' + process.env['INFURA_KEY'],
-        enabled: process.env['SHOULD_FORK'] === '1',
-      },
-    },
-    local_development: {
-      url: 'http://127.0.0.1:7545',
-    },
-    kovan: {
-      url: 'https://kovan.infura.io/v3/' + process.env['INFURA_KEY'],
-      accounts: process.env['DEVNET_PRIVKEY']
-        ? [process.env['DEVNET_PRIVKEY']]
-        : [],
-    },
-    mainnet: {
-      url: 'https://mainnet.infura.io/v3/' + process.env['INFURA_KEY'],
-      accounts: process.env['MAINNET_PRIVKEY']
-        ? [process.env['MAINNET_PRIVKEY']]
-        : [],
-    },
-    fork: {
-      url: 'http://127.0.0.1:8545/',
-    },
-    arbitrum1: {
-      url: 'https://arb1.arbitrum.io/rpc',
-      accounts: process.env['MAINNET_PRIVKEY']
-        ? [process.env['MAINNET_PRIVKEY']]
-        : [],
-    },
-    nova: {
-      url: 'https://nova.arbitrum.io/rpc',
-      accounts: process.env['MAINNET_PRIVKEY']
-        ? [process.env['MAINNET_PRIVKEY']]
-        : [],
-    },
-    rinkeby: {
-      url: 'https://rinkeby.infura.io/v3/' + process.env['INFURA_KEY'],
-      accounts: process.env['DEVNET_PRIVKEY']
-        ? [process.env['DEVNET_PRIVKEY']]
-        : [],
-    },
-    arbRinkeby: {
-      url: 'https://rinkeby.arbitrum.io/rpc',
-      accounts: process.env['DEVNET_PRIVKEY']
-        ? [process.env['DEVNET_PRIVKEY']]
-        : [],
-    },
-    goerli: {
-      url: 'https://goerli.infura.io/v3/' + process.env['INFURA_KEY'],
-      accounts: process.env['DEVNET_PRIVKEY']
-        ? [process.env['DEVNET_PRIVKEY']]
-        : [],
-    },
-    arbGoerliRollup: {
-      url: 'https://goerli-rollup.arbitrum.io/rpc',
-      accounts: process.env['DEVNET_PRIVKEY']
-        ? [process.env['DEVNET_PRIVKEY']]
-        : [],
-    },
-    arbitrum: {
-      url: 'http://127.0.0.1:8547',
-      gas: 999999999999999,
-      accounts: {
-        mnemonic:
-          'jar deny prosper gasp flush glass core corn alarm treat leg smart',
-        path: "m/44'/60'/0'/0",
-        initialIndex: 0,
-        count: 10,
-      },
-      timeout: 100000,
-    },
-  },
-  typechain: {
-    outDir: 'build/types',
-    target: 'ethers-v5',
-  },
-  spdxLicenseIdentifier: {
-    overwrite: false,
-    runOnCompile: true,
-  },
-  gasReporter: {
-    currency: 'USD',
-    gasPrice: 20,
-    enabled: process.env.REPORT_GAS ? true : false,
-    maxMethodDiff: 5,
-  },
-  namedAccounts: {
-    deployer: {
-      default: 0,
-    },
-  },
-  mocha: {
-    timeout: 0,
-    bail: true,
-  },
-  etherscan: {
-    apiKey: {
-      mainnet: process.env['ETHERSCAN_API_KEY'],
-      kovan: process.env['ETHERSCAN_API_KEY'],
-      rinkeby: process.env['ETHERSCAN_API_KEY'],
-      goerli: process.env['ETHERSCAN_API_KEY'],
-      arbitrumOne: process.env['ARBISCAN_API_KEY'],
-      arbitrumTestnet: process.env['ARBISCAN_API_KEY'],
-      nova: '0',
-      arbGoerliRollup: '0',
-    },
-    customChains: [
-      {
-        network: 'nova',
-        chainId: 42170,
-        urls: {
-          apiURL: 'https://nova-explorer.arbitrum.io/api',
-          browserURL: 'https://nova-explorer.arbitrum.io/',
-        },
-      },
-      {
-        network: 'arbGoerliRollup',
-        chainId: 421613,
-        urls: {
-          apiURL: 'https://goerli-rollup-explorer.arbitrum.io/api',
-          browserURL: 'https://goerli-rollup-explorer.arbitrum.io/',
-        },
-      },
-    ],
-  },
-}
-
-if (process.env['RINKEBY_URL'] && process.env['RINKEBY_MNEMONIC']) {
-  ;(config.networks as any)['rinkeby'] = {
-    url: process.env['RINKEBY_URL'] || '',
-    accounts: [process.env['RINKEBY_MNEMONIC'] || ''],
-    network_id: 4,
-    confirmations: 1,
-  }
-}
-
-if (process.env['ROPSTEN_URL'] && process.env['ROPSTEN_MNEMONIC']) {
-  ;(config.networks as any)['ropsten'] = {
-    url: process.env['ROPSTEN_URL'] || '',
-    accounts: [process.env['ROPSTEN_MNEMONIC'] || ''],
-    network_id: 3,
-    confirmations: 1,
-  }
-}
-
-if (process.env['KOVAN_URL'] && process.env['KOVAN_MNEMONIC']) {
-  ;(config.networks as any)['kovan'] = {
-    url: process.env['KOVAN_URL'] || '',
-    accounts: [process.env['KOVAN_MNEMONIC'] || ''],
-    network_id: 42,
-    confirmations: 4,
-  }
-}
-
-if (process.env['GOERLI_URL'] && process.env['GOERLI_MNEMONIC']) {
-  ;(config.networks as any)['goerli'] = {
-    url: process.env['GOERLI_URL'] || '',
-    accounts: [process.env['GOERLI_MNEMONIC'] || ''],
-    network_id: 42,
-    confirmations: 4,
-  }
-}
-
-if (!process.env['DEVNET_PRIVKEY']) console.warn('No devnet privkey set')
-
-export { config }
